+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **command:** `Command` is now typed based on the execute function
 - **command:** `Command` `canExecute` now support `boolean`
 - **command:** `SsvCommand/SsvCommandRef` is now typed
+- **command:** add `CommandInput` type
 
 ### BREAKING CHANGES